refactor(riichi-city): move round advance into the riichi_city module
round_advance_watcher and its round_advance_delay are Riichi City-only
(Majsoul and Tenhou have no per-round ready step), so they don't belong
in the platform-agnostic autoplay manager. Move them, the RC-only
enabled gate, and their test into a new autoplay::riichi_city::round_advance
submodule; the manager just spawns the watcher.
