+++ src/autoplay/riichi_city/mod.rs
@@ -9,6 +9,8 @@ use crate::bridge::riichi_city::build;
 use crate::config::DelayMode;
 use crate::schema::MjaiEvent;
 
+pub mod round_advance;
+
 #[derive(Default)]
 pub struct RiichiCityAutoplay;
